feat: Add portal support
The portal config receives either a bool or a dictionary of function,
when portal is True, a portal.py will be required in the current directory and automatically loaded.
The functions defined either in portal.py or as a dict of function can now be reference by name and ke
y, along different stages in the crawling/scraping process.
- Remove automatic cleanup from .start method
- Add .end method for manual cleanup
- Update .data method to with parameter to conditionally dump data to file
- Config output format transform, now expects a function name defined in the portal
- Custom utils can be defined in the portal and reference by name e.g | subract 1 | custom_util